importer/requirements.txt
openpyxl==2.2.5
requests==2.7.0
python-slugify==1.1.3
Jinja2==2.7.3
